backend/requirements.txt
fastapi
uvicorn
firebase-admin
pydantic
python-dotenv
pytest
httpx
python-multipart
google-generativeai
python-docx
python-pptx
markdown2
